cache.c: replaced CDS with dir flock, tackled critical sections
This replaces DB_INIT_CDB mode (aka Concurrent Data Store) with custom
per-cache locking.  I think that CDS is complicated and somewhat buggy:
the isalive() mechanism based on kill/ESRCH is unreliable, especially
since Linux has only 32K pids.  And CDS uses global per-database lock
anyway.

This also elaborates on the subject of critical sections.  The idea is
that some steps must be taken, as a means of protection, before passing
control to the berkeley db library.  This includes locking for
reader/write mutual exclusion, and blocking signals so that writes can
be complete.  Note that reads apparently can trigger writes due to
memory pool flushes.  So the signals should be blocked even for reads.

Also note that we have to assume that system calls, notably flock,
should never fail, provided that they are called with the right
arguments.  And wrong arguments, excluding outright bug in the program,
are only possible due to stack and/or heap corruption, and that bad a
problem cannot be handled gracefully anyway.  Linux flock(2) manpage
lists ENOLCK as possible return for kernel allocation issues.  But
note that flocks are obtained on file table entries, and thus it is
not possible to obtain more flocks that file descriptors.  And the
latter are subject for rlimit.  There is no good reason why the kernel
could run out of locks.