cache.c: don't close db after fork
It is not yet clear to me if the cache can survive fork and be used in
child processes (this will not work with berkeley db locking mechanism
which identifies lockers, but we use our own fcntl-based locking
instead).  It is clear, though, that double-close after fork leads to
"environment reference count went negative" error messages.

So, this is probably not the best design decision, which is to check the
pid in lower-level implementation routine, but at least this does not
seem wrong.