db.c: switch from DB_HASH to DB_BTREE
We use this library to process rpm packages; the keys we use start
with rpm basename.  When rpms are scanned in alphabetical order,
DB_BTREE can give a big advantage due to consistent read-ahead.

There's no need to pretend that something more general is going on
here; we don't have big keys which need to be shortened with SHA1.