new configuration with backends
The qa-cache backend is nice because it provides easy naming due
to filesystem, i.e. ~/.qa-cache/name, and so rpmcache_open("name")
is easily possible.  This is not so with other backends, there is
no direct mapping of a "name" to the corresponding memcached port.

The idea is then to add another level of indirection, such as this:

$ cat rpmcache.conf
rpmhdrcache memcached --socket="/tmp/memcached-socket"

which means that "rpmhdrcache" is served by memcached with certain
configuration.

Support for backends is now moved into rpmcache.c.  This greatly
simplifies the code in hdrcache.c (and greatly complicates the
the code in rpmcache.c).